Refactor bot to structure the bot in a way that facilitates future development instead of just being a script.
